requirements.txt
Flask
Flask-SQLAlchemy
APScheduler
requests
beautifulsoup4
python-dotenv
psycopg2-binary
